[#13] refactor : getSearchPosts 함수를 hook로 만들었습니다
getSearchPosts함수와 포함되는 state를 모두 useSearchPosts 라는훅으로
만들어서 추상화와 재사용성을 높였습니다.
