+++ hooks/useSearchPosts.ts
@@ -1,0 +1,27 @@
+import { useState } from "react";
+
+export default function useSearchPosts(allPosts) {
+  const [searchTitle, setSearchTitle] = useState("");
+  const [searchPosts, setSearchPosts] = useState([]);
+
+  const onChangeSearchTitle = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setSearchTitle(e.target.value);
+  };
+
+  const getSearchPosts = () => {
+    if (searchTitle !== "") {
+      const tmp = [];
+      allPosts.map((post) => {
+        const regex = new RegExp(searchTitle, "gim");
+        if (regex.test(post.title)) {
+          tmp.push(post);
+        }
+      });
+      setSearchPosts(tmp);
+    } else {
+      setSearchPosts([]);
+    }
+  };
+
+  return { searchTitle, searchPosts, onChangeSearchTitle, getSearchPosts };
+}
